pyproject.toml
[tool.poetry]
name = "jenkins-decryptor"
version = "0.1.0"
description = ""
authors = ["dadevel <dadevel@disroot.org>"]
license = "MIT"
packages = [{include = "jenkinsdecryptor"}]
readme = "README.md"

[tool.poetry.scripts]
jenkins-decryptor = "jenkinsdecryptor.main:main"

[tool.poetry.dependencies]
python = "^3.12"
pycryptodome = "^3.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
